Former-commit-id: 5294c7c2ef871058c95891650ebb4944ae3ce7a3

+++ markov.py
@@ -40,8 +40,8 @@ class Markov:
         Returns:
             transition probabilities
         """
-        self.transitions = defaultdict(lambda: defaultdict(int))
-        self.emissions = defaultdict(lambda: defaultdict(int))
+        self.transitions = defaultdict(lambda: defaultdict(float))
+        self.emissions = defaultdict(lambda: defaultdict(float))
 
         for corpus_entry in corpus:
             if corpus_entry[0] is None:
